Backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
python-dotenv>=1.0.0
pypdf>=4.0.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0
sentence-transformers>=2.2.0
huggingface-hub>=0.20.0
torch>=2.0.0
openai>=1.10.0